requirements.txt
streamlit>=1.32
python-dotenv>=1.0.1
requests>=2.31
PyPDF2>=3.0.0
python-docx>=0.8.11
wikipedia>=1.4.0
rapidfuzz>=3.9.6
